Removing GATB and previous denovo code from codebase
